Work around for class loader conflict in Jenkins
Found this reference:
    http://jenkins.361315.n4.nabble.com/ClassLoader-in-plugins-td1470791.html

I thought the problem was related to clustering; I would add another
Vert.x instance to the cluster, send some messages, remove it, add it
back and then the messages published by the listeners would cause
exceptions in the EventBus code.  The cluster join/part/join may have
been a red herring, as this seems to solve the problem.